updated both index.html and styles.css
I did a lot to finish this project out. I put the main image in a div so I can create a mask that would activate on hover (I had to look up how to do this), I also changed out the h2 element to a h1 for SEO reasons. The main css updates was to show the active states on the heading, main image, and creator name. Looks like I will have to adjust some of the padding surrounding the text a little bit to better match the mock-up